requirements.txt
fastapi[all]
loguru
asyncio
langdetect
pillow
opencv-python==3.4.17.61
imutils